Remove states sent in the projects' board data
Externalise the state mapping
